requirements.txt
langchain>=0.3.0
langchain-classic
langchain-google-genai
langchain-core>=0.3.0
pandas>=2.0.0
python-dotenv>=1.0
pydantic>=2.0
streamlit>=1.28.0
